Fix streamsubscription no longer receiving events without any form of notification
If the socketEndPoint got closed the listenerTask would stop running. However, the exception this caused was eaten up by the executor and the streamsubscription would happily try to read incoming data until the end of time. By stopping the handlers in this scenario the steamsubscription receives an exception and can act on the change in connectivity.